tests/wpt/web-platform-tests/notifications/notification-constructor-basic.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Web Notifications Test: notification - constructor (basic)</title>
<link rel="author" title="Intel" href="http://www.intel.com/">
<link rel="author" title="Xin Liu" href="mailto:xinx.liu@intel.com">
<link rel="help" title="5 API" href="http://www.w3.org/TR/notifications/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
  <div id="log"></div>
  <script>
    test(function() {
      var notification = new Notification("New Email Received", {
        body: "Room 101"
      });
      assert_equals(typeof notification, "object");
    }, "Test checks that notification constructor is correct");
  </script>
</body>
